requirements.txt
openrgb-python
requests
